refactor code and route location
