+++ src/Root.js
@@ -2,13 +2,12 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
-import { MODULE_ROOT_ROUTE, PERMISSION_SET_ROUTES } from './constants';
+import { MODULE_ROOT_ROUTE } from './constants';
 import { ConsortiumManagerContextProvider } from './contexts';
 import { eventHandler } from './eventHandler';
 import { ConsortiumManager } from './routes';
 import ConsortiumSettings from './settings';
 import { checkConsortiumAffiliations } from './utils';
-import { PermissionSetsCompare } from './routes/ConsortiumManager/settings/users';
 
 class Root extends React.Component {
   // Checks whether to show "Switch active affiliation" action item in the profile dropdown (package.json::stripes.links.userDropdown[0])
@@ -36,10 +35,6 @@ class Root extends React.Component {
     return (
       <ConsortiumManagerContextProvider>
         <Switch>
-          <Route
-            path={PERMISSION_SET_ROUTES.COMPARE}
-            component={PermissionSetsCompare}
-          />
           <Route
             path={MODULE_ROOT_ROUTE}
             component={ConsortiumManager}
