+++ src/routes/ConsortiumManager/settings/users/general/PermissionSets/constants.js
@@ -1,0 +1,4 @@
+export const PERMISSION_SET_ROUTES = {
+  COMPARE: '/consortia-settings/users/perms/compare',
+  PERMISSION_SETS: '/consortia-settings/users/perms',
+};
