+++ src/routes/ConsortiumManager/settings/users/general/PermissionSets/index.js
@@ -1,2 +1,1 @@
 export { PermissionSets } from './PermissionSets';
-export { PermissionSetsCompare } from './PermissionSetsCompare/index';
