+++ src/routes/ConsortiumManager/settings/users/index.js
@@ -1,2 +1,1 @@
 export { default } from './UsersSettings';
-export { PermissionSetsCompare } from './general/PermissionSets';
